+++ src/mate/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+0.2
+---
+
+ * Add `StopCommand` to stop a running server
+ * Add `--force-keep-alive` option to `ServeCommand` to restart server if it was stopped
+
 0.1
 ---